Player can now attack and do damage
Added hitboxes for the enemies and hitboxes for the light attacks (new object for light attack hitbox) (new file for enemy health. The file for the hitbox of the attack is called L (I had naming issues earlier, but I can change it back when im less tired). Added rigid body to enemy so that it could register the collision. TLDR the player can now kill the enemy.